auth-app/chatbot-api/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
qdrant-client
python-multipart
httpx
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
email-validator